refactor(jukebox): Remove unused track file
